+++ packages/react-async/src/Prefetcher.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import {Omit} from '@shopify/useful-types';
 
 import {PrefetchContext, PrefetchManager} from './context/prefetch';
 import {EventListener} from './EventListener';