+++ packages/react-html/src/components/HtmlAttributes.tsx
@@ -1,8 +1,6 @@
-import {FirstArgument} from '@shopify/useful-types';
-
 import {useHtmlAttributes} from '../hooks';
 
-type Props = FirstArgument<typeof useHtmlAttributes>;
+type Props = Parameters<typeof useHtmlAttributes>[0];
 
 export function HtmlAttributes(props: Props) {
   useHtmlAttributes(props);